Added permissions, fixed a ton of bugs

+++ database/connection.php
@@ -1,5 +1,5 @@
 <?php 
-$db = new PDO('sqlite:'.$_SERVER['DOCUMENT_ROOT'].'/database/todo.db');
+$db = new PDO('sqlite:database/todo.db');
 $db->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
 $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 $db->exec('PRAGMA foreign_keys = ON;');
@@ -14,6 +14,31 @@ function isLoginCorrect($username, $password) {
   return ($user['name']!== false && sha1($password)==$user['passwordHash']);
 }
 
+function checkOwnership($todoPageID) {
+  global $db;
+  
+  $username = null;
+
+  if(!empty($_SESSION['username'])){
+    $username = $_SESSION['username'];
+  }
+
+  $stmt = $db->prepare("SELECT *
+                        FROM TodoLists
+                        WHERE ID=?");
+  $stmt->execute(array($todoPageID));
+
+  $todoPage = $stmt->fetch(); 
+
+  if($todoPage['public'] == '1'){
+    return true;
+  }else if(!empty($username) && $todoPage['user'] == $username){
+    return true;
+  }else{
+    return false;
+  }
+}
+
 function createUser($username, $password, $name) {
   global $db;  
 
@@ -24,34 +49,53 @@ function createUser($username, $password, $name) {
 }
 
 function addTodoItemToList($content, $todoPage) {
-  global $db;  
+  global $db;
+
+  if (!checkOwnership($todoPage)){
+    return false;
+  }
   
   date_default_timezone_set('Europe/Lisbon');
   $date = date('h:i d/m/Y', time());
   $stmt = $db->prepare('INSERT INTO TodoItems VALUES (NULL,?, ?, ?, "notdone")');
   $stmt->execute(array($content, $date, $todoPage));
+
+  return true;
 }
 
 
-function addTodoList($username, $title) {
+function addTodoList($username, $title, $public) {
   global $db;  
   
-  $stmt = $db->prepare('INSERT INTO TodoLists VALUES (NULL, ?, ?, "1")');
-  $stmt->execute(array($title, $username));
+  $stmt = $db->prepare('INSERT INTO TodoLists VALUES (NULL, ?, ?, ?)');
+  $stmt->execute(array($title, $username, $public));
 }
 
 function removeTodoItem($ID) {
     global $db;
 
+    $todoItem = getTodoItemByID($ID);
+    if (!checkOwnership($todoItem['todoList'])){
+      return false;
+    }
+
     $stmt = $db->prepare('DELETE FROM TodoItems WHERE ID=?');
     $stmt->execute(array($ID));
+    
+    return true;
 }
 
 function removeTodoList($ID) {
   global $db;
 
+  if (!checkOwnership($ID)){
+    return false;
+  }
+
   $stmt = $db->prepare('DELETE FROM TodoLists WHERE ID=?');
   $stmt->execute(array($ID));
+
+  return true;
 }
 
 function editUser($old_username, $username, $password, $name) {
@@ -68,39 +112,73 @@ function editUser($old_username, $username, $password, $name) {
   $_SESSION['username'] = $username;  
 }
 
-function editItem($content, $todoItem) {
+function editItem($content, $todoItemID) {
   global $db;
+
+  $todoItem = getTodoItemByID($todoItemID);
+
+  if (!checkOwnership($todoItem['todoList'])){
+    return false;
+  }
+
   $stmt = $db->prepare('UPDATE TodoItems 
                         SET content=?
                         WHERE ID=?');
-  $stmt->execute(array($content,$todoItem));
+  $stmt->execute(array($content,$todoItemID));
+
+  return true;
  
 }
 
-function editstatusItem($status, $todoItem) {
+function editstatusItem($status, $todoItemID) {
   global $db;
+
+  $todoItem = getTodoItemByID($todoItemID);
+  
+  if (!checkOwnership($todoItem['todoList'])){
+    return false;
+  }
+
   $stmt = $db->prepare('UPDATE TodoItems 
                         SET status=?
                         WHERE ID=?');
-  $stmt->execute(array($status,$todoItem)); 
+  $stmt->execute(array($status,$todoItemID)); 
+
+  return true;
 }
 
 
-function getTodoItemByID($todoItem) {
+function getTodoItemByID($todoItemID) {
   global $db;
+
   $stmt = $db->prepare("SELECT *
                         FROM TodoItems
                         WHERE ID=?");
-  $stmt->execute(array($todoItem));
-  return $stmt->fetch();
+  $stmt->execute(array($todoItemID));
+
+  $todoItem = $stmt->fetch();
+
+  if (!checkOwnership($todoItem['todoList'])){
+    return false;
+  }
+
+  return $todoItem;
 }
 
 function getLastTodoItem(){
   global $db;
+
   $stmt = $db->prepare("SELECT *
                         FROM TodoItems
                         ORDER BY ID DESC");
-  $stmt->execute(array($todoItem));
-  return $stmt->fetch();
+  $stmt->execute();
+  
+  $todoItem = $stmt->fetch();
+  
+  if (!checkOwnership($todoItem['todoList'])){
+    return false;
+  }
+  
+  return $todoItem;
 }
 ?>
